Add Pengaturan screen with user profile and settings functionality
